Implemented! (kind of)
We got a full-on implementation going. The only thing missing is the
tension force in the rod between the pendula. Once we calculate that, we
can drop it into a function we wrote, and we will have a completed
model. We also plan to create an animation to help us visualize the
system.