More Graphs, More Computations
We made some more graphs and got a pretty good definition of chaos in
counting the peaks of the initial vs. current position curve. We did a
bunch of computation-intensive scripts and functions and made some new
graphs. Turns out chaos increases exponentially by this definition. Some
of the data is still a little suspicious-looking, though...